requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
defusedxml==0.7.1
dnspython==2.4.2
docker==6.1.3
ecdsa==0.18.0
exceptiongroup==1.1.2
exchangelib==5.1.0
fastapi==0.97.0
fonttools==4.42.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.0
oauthlib==3.2.2
packaging==23.1
passlib==1.7.4
Pillow==10.0.1
pluggy==1.2.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.7.0
pyparsing==3.1.1
pyspnego==0.9.2
pytest==7.4.0
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
pywin32==306
requests==2.31.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.15
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
uvicorn==0.23.2
websocket-client==1.6.